refactor: cleaning up code
