+++ lib/features/auth/models/user/user_preferences.dart
@@ -2,6 +2,7 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:isar/isar.dart';
 
+import '../../../../core/core.dart';
 import '../../../core/enums.dart';
 
 part 'user_preferences.g.dart';
