+++ lib/features/auth/providers/user/user_notifier.dart
@@ -4,6 +4,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:fpdart/fpdart.dart';
 import 'package:sickler/repositories/user/user_repository.dart';
 
+import '../../../../core/core.dart';
+import '../../auth.dart';
 import '../../core/failure.dart';
 import '../../models/models.dart';
