+++ lib/features/home/screens/components/meds_reminder_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 
-import '../../../core/core.dart';
+
 
 class MedsReminderItem extends StatelessWidget {
   const MedsReminderItem({super.key});
