+++ lib/features/meds/screens/components/meds_history_item.dart
@@ -3,7 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:gap/gap.dart';
-import '../../../core/core.dart';
+
+import '../../../../../core/core.dart';
 
 class MedsHistoryItem extends StatelessWidget {
   const MedsHistoryItem({super.key, this.mode = MedsHistoryMode.daily});
