+++ lib/features/meds/screens/components/meds_reminder_card.dart
@@ -3,10 +3,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:gap/gap.dart';
-import 'package:sickler/screens/meds/meds_details_screen.dart';
 
-import '../../../core/core.dart';
-import '../../global_components/components.dart';
+import '../../../../components/components.dart';
+import '../../../../../core/core.dart';
+import '../../meds.dart';
 
 class MedsScreenReminderCard extends StatefulWidget {
   const MedsScreenReminderCard({
