+++ lib/features/meds/screens/components/sickler_calendar_day_selector_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 
-import '../../../core/core.dart';
+import '../../../../../core/core.dart';
 
 class SicklerCalendarDaySelectorItem extends StatefulWidget {
   const SicklerCalendarDaySelectorItem({
