+++ lib/features/profile/screens/components/app_radio.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../core/core.dart';
+import '../../../../../core/core.dart';
