+++ lib/features/profile/screens/components/genotype_selector.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:gap/gap.dart';
 
-import '../../../../core/core.dart';
+import '../../../../../core/core.dart';
 import 'components.dart';
 
 class GenotypeSelector extends StatefulWidget {
