+++ lib/features/profile/screens/components/genotype_selector_item.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../core/core.dart';
+import '../../../../../core/core.dart';
 
 
 class GenotypeSelectorItem extends StatefulWidget {
