+++ lib/features/water/models/water_preferences.dart
@@ -1,6 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:isar/isar.dart';
 
+import '../../../core/core.dart';
 import '../../core/enums.dart';
 
 part 'water_preferences.g.dart';
