+++ lib/features/water/models/water_stats.dart
@@ -2,6 +2,7 @@ import 'package:equatable/equatable.dart';
 import 'package:sickler/models/water/water_log.dart';
 
 import '../../core/core.dart';
+import '../water.dart';
 
 class WaterStats extends Equatable {
   ///---Today---///
