+++ lib/features/water/screens/components/water_volume_selector.dart
@@ -2,7 +2,6 @@ import 'package:fluentui_system_icons/fluentui_system_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-import '../../global_components/selector_item.dart';
 
 class WaterVolumeSelector extends StatefulWidget {
   const WaterVolumeSelector({super.key, required this.selectedVolume});
