+++ lib/features/water/screens/water_screen.dart
@@ -6,19 +6,15 @@ import 'package:gap/gap.dart';
 import 'package:sickler/core/constants.dart';
 import 'package:sickler/core/core.dart';
 import 'package:sickler/core/snackbar_notifier.dart';
-import 'package:sickler/models/water/water_log.dart';
-import 'package:sickler/models/water/water_preferences.dart';
-import 'package:sickler/features/auth/providers/auth_providers.dart';
-import 'package:sickler/screens/global_components/components.dart';
-import 'package:sickler/screens/water/charts/chart_data_transformer.dart';
-import 'package:sickler/screens/water/charts/line_chart_widget.dart';
-import 'package:sickler/screens/water/components/water_volume_selector.dart';
 
-import '../../models/models.dart';
-import '../../features/water/providers/water_log_notifier.dart';
-import '../../providers/water/water_providers.dart';
+import '../../../components/components.dart';
+import '../../auth/auth.dart';
+import '../water.dart';
 import 'charts/bar_chart_widget.dart';
+import 'charts/chart_data_transformer.dart';
+import 'charts/line_chart_widget.dart';
 import 'components/components.dart';
+import 'components/water_volume_selector.dart';
 
 class WaterScreen extends ConsumerStatefulWidget {
   static const String id = "water";
