+++ test/providers/auth/auth_notifier_provider_test.dart
@@ -3,6 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:fpdart/fpdart.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:sickler/core/core.dart';
+import 'package:sickler/features/auth/auth.dart';
 import 'package:sickler/models/models.dart';
 import 'package:sickler/features/auth/providers/auth/auth_notifier.dart';
 import 'package:sickler/features/auth/repositories/auth/auth_repository.dart';
