+++ test/repositories/auth/auth_repository_test.dart
@@ -4,9 +4,9 @@ import 'package:fpdart/fpdart.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:sickler/core/core.dart';
 import 'package:sickler/features/auth/models/user/app_user.dart';
-import 'package:sickler/features/auth/providers/auth_providers.dart';
-import 'package:sickler/repositories/repositories.dart';
-import 'package:sickler/services/services.dart';
+import 'package:sickler/features/auth/services/auth/remote/auth_service.dart';
+import 'package:sickler/features/auth/services/user/local/user_local_service.dart';
+
 
 class MockAuthService extends Mock implements AuthService {}
